Move shadowsocks config tests to `shadowsocks_test.go` and clean up.
